requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.4.0